Update patch set 2
Patch Set 2:

Couldn't find online what it does, the T810 stock firmware is at least shipped with the library. When I do a diff with the gts210ltexx version they differ (could be a date difference, size is the same)

Will test what the build for gts210wifi does when I leave it out.

Patch-set: 2