Update patch set 2
Patch Set 2:

What I'm up to is the fact that Jan Huijsmans deleted mediacodec.te for a reason in that old commit. I'm not trying to argue, just reporting what I've demonstrated to myself on my own tablet. Whether Netflix streams a video or not is kind of obvious. I'm running another build right now with the official gts2-common repo so I'll check again.

Patch-set: 2